extract progress bar functions to subscription view helper
