+++ apps/concierge_site/lib/views/commuter_rail_subscription_view.ex
@@ -1,6 +1,7 @@
 defmodule ConciergeSite.CommuterRailSubscriptionView do
   use ConciergeSite.Web, :view
-  import ConciergeSite.SubscriptionViewHelper, only: [atomize_keys: 1, travel_time_options: 0]
+  import ConciergeSite.SubscriptionViewHelper,
+    only: [atomize_keys: 1, travel_time_options: 0, progress_link_class: 3]
   alias AlertProcessor.Model.Trip
 
   @type trip_type :: :one_way | :round_trip
@@ -9,28 +10,9 @@ defmodule ConciergeSite.CommuterRailSubscriptionView do
   train: [:train, :preferences],
   preferences: [:preferences]}
 
-  @doc """
-  Provide css class to disable links within the subscription flow progress
-  bar
-  """
-  def progress_link_class(:trip_type, _step), do: "disabled-progress-link"
+  defdelegate progress_step_classes(page, step), to: ConciergeSite.SubscriptionViewHelper
 
-  def progress_link_class(page, step) do
-    if @disabled_progress_bar_links |> Map.get(page) |> Enum.member?(step) do
-      "disabled-progress-link"
-    end
-  end
-
-  @doc """
-  Provide css classes for the text and circle in progress bar steps
-  """
-  def progress_step_classes(page, step) when (page == step) do
-    %{circle: "active-circle", name: "active-page"}
-  end
-
-  def progress_step_classes(_page, _step) do
-    %{}
-  end
+  def progress_link_class(page, step), do: progress_link_class(page, step, @disabled_progress_bar_links)
 
   @doc """
   Provide description text for Trip Info page based on which trip type selected
