+++ apps/concierge_site/lib/views/ferry_subscription_view.ex
@@ -1,30 +1,13 @@
 defmodule ConciergeSite.FerrySubscriptionView do
   use ConciergeSite.Web, :view
+  import ConciergeSite.SubscriptionViewHelper,
+    only: [progress_link_class: 3]
 
   @disabled_progress_bar_links %{trip_info: [:trip_info, :ferry, :preferences],
   ferry: [:ferry, :preferences],
   preferences: [:preferences]}
 
-  @doc """
-  Provide css class to disable links within the subscription flow progress
-  bar
-  """
-  def progress_link_class(:trip_type, _step), do: "disabled-progress-link"
+  defdelegate progress_step_classes(page, step), to: ConciergeSite.SubscriptionViewHelper
 
-  def progress_link_class(page, step) do
-    if @disabled_progress_bar_links |> Map.get(page) |> Enum.member?(step) do
-      "disabled-progress-link"
-    end
-  end
-
-  @doc """
-  Provide css classes for the text and circle in progress bar steps
-  """
-  def progress_step_classes(page, step) when (page == step) do
-    %{circle: "active-circle", name: "active-page"}
-  end
-
-  def progress_step_classes(_page, _step) do
-    %{}
-  end
+  def progress_link_class(page, step), do: progress_link_class(page, step, @disabled_progress_bar_links)
 end
