+++ apps/concierge_site/lib/views/subway_subscription_view.ex
@@ -2,7 +2,7 @@ defmodule ConciergeSite.SubwaySubscriptionView do
   use ConciergeSite.Web, :view
   import ConciergeSite.SubscriptionViewHelper,
     only: [joined_day_list: 1, travel_time_options: 0, format_time: 1,
-           time_option_local_strftime: 1, atomize_keys: 1]
+           time_option_local_strftime: 1, atomize_keys: 1, progress_link_class: 3]
 
   @typedoc """
   Possible values for trip types in Create Subscription flow
@@ -12,28 +12,9 @@ defmodule ConciergeSite.SubwaySubscriptionView do
   @disabled_progress_bar_links %{trip_info: [:trip_info, :preferences],
     preferences: [:preferences]}
 
-  @doc """
-  Provide css class to disable links within the subscription flow progress
-  bar
-  """
-  def progress_link_class(:trip_type, _step), do: "disabled-progress-link"
+  defdelegate progress_step_classes(page, step), to: ConciergeSite.SubscriptionViewHelper
 
-  def progress_link_class(page, step) do
-    if @disabled_progress_bar_links |> Map.get(page) |> Enum.member?(step) do
-      "disabled-progress-link"
-    end
-  end
-
-  @doc """
-  Provide css classes for the text and circle in progress bar steps
-  """
-  def progress_step_classes(page, step) when (page == step) do
-    %{circle: "active-circle", name: "active-page"}
-  end
-
-  def progress_step_classes(_page, _step) do
-    %{}
-  end
+  def progress_link_class(page, step), do: progress_link_class(page, step, @disabled_progress_bar_links)
 
   @doc """
   Provide description text for Trip Info page based on which trip type selected
